Bring back doFmtVerbLevelColor on Windows
